backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.0
numpy==1.24.3
pillow==10.0.0
python-multipart==0.0.6
joblib==1.3.2
scipy==1.11.0
opencv-python==4.8.1